requirements1.txt
# Minimal Requirements - Let pip resolve compatible versions
streamlit>=1.28.0
langchain>=0.1.0
langchain-google-genai>=1.0.0
langchain-huggingface>=0.0.3
langchain-community>=0.0.30
faiss-cpu>=1.7.0
python-dotenv>=1.0.0
PyPDF2>=3.0.0

# Speech Recognition (Optional)
SpeechRecognition>=3.10.0
gTTS>=2.4.0

# Core ML libraries
sentence-transformers>=2.2.0
transformers>=4.35.0
numpy>=1.24.0